+++ lib/src/deriv_chart/interactive_layer/interactable_drawings/horizontal_line/horizontal_line_interactable_drawing.dart
@@ -8,13 +8,12 @@ import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_too
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/models/animation_info.dart';
 import 'package:deriv_chart/src/deriv_chart/interactive_layer/interactable_drawings/drawing_adding_preview.dart';
 import 'package:deriv_chart/src/deriv_chart/interactive_layer/interactable_drawings/horizontal_line/horizontal_line_adding_preview_desktop.dart';
-import 'package:deriv_chart/src/widgets/color_picker/color_picker_dropdown_button.dart';
 import 'package:deriv_chart/src/models/axis_range.dart';
 import 'package:deriv_chart/src/models/chart_config.dart';
 import 'package:deriv_chart/src/theme/chart_theme.dart';
-import 'package:deriv_chart/src/theme/design_tokens/core_design_tokens.dart';
 import 'package:deriv_chart/src/theme/painting_styles/line_style.dart';
-import 'package:deriv_chart/src/widgets/line_thickness/line_thickness_dropdown_button.dart';
+import 'package:deriv_chart/src/widgets/color_picker/color_picker_dropdown_button.dart';
+import 'package:deriv_chart/src/widgets/dropdown/line_thickness/line_thickness_dropdown_button.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
@@ -291,32 +290,4 @@ class HorizontalLineInteractableDrawing
           ));
         },
       );
-
-// Widget _buildLineThicknessIcon() => SizedBox(
-//       width: 32,
-//       height: 32,
-//       child: TextButton(
-//         style: TextButton.styleFrom(
-//           foregroundColor: Colors.white38,
-//           alignment: Alignment.center,
-//           padding: const EdgeInsets.symmetric(),
-//           shape: RoundedRectangleBorder(
-//             borderRadius: BorderRadius.circular(4),
-//           ),
-//         ),
-//         onPressed: () {
-//           // update line thickness
-//         },
-//         child: Text(
-//           '${config.lineStyle.thickness.toInt()}px',
-//           style: const TextStyle(
-//             fontSize: 14,
-//             color: CoreDesignTokens.coreColorSolidSlate50,
-//             fontWeight: FontWeight.normal,
-//             height: 2,
-//           ),
-//           textAlign: TextAlign.center,
-//         ),
-//       ),
-//     );
 }
